requirements.txt
Django==5.2.3
pillow==10.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
gradio==4.31.0
gradio-client==0.17.0